Add files via upload
Markov Decision Process (MDP) is a mathematical framework used to model decision-making in situations where an agent interacts with an environment over a series of discrete time steps. MDPs are a fundamental concept in reinforcement learning and provide a formal structure for studying and solving problems related to decision making under uncertainty.

In an MDP, the agent selects actions in various states, and the environment responds by transitioning the system to new states, providing rewards based on these actions. The key components of an MDP include states, actions, transition probabilities, rewards, and policies. MDPs are built on the Markov property, ensuring that the future state depends only on the current state and action, simplifying the modeling process.

Value functions, such as the state-value and action-value functions, are fundamental to MDPs. These functions estimate the expected cumulative future rewards and guide the agent in making decisions. The Bellman equations capture the recursive relationships between values of states or state-action pairs, facilitating the computation of optimal policies.

The pursuit of an optimal policy, one that maximizes cumulative rewards, is a central goal in MDPs. Algorithms like Q-learning and policy iteration are employed to find these optimal policies efficiently. MDPs have widespread applications in AI, ranging from robotics and autonomous systems to game playing and decision support systems. They provide a robust framework for modeling complex decision-making scenarios and have proven instrumental in advancing AI capabilities in dynamic and uncertain environments.